discord-exporter/requirements.txt
discord.py
pandas
openpyxl
gspread
google-auth-oauthlib
google-auth-httplib2
